code-review-agent/requirements.txt
crewai
bitbucket-pipes-toolkit==4.*
PyYAML
tiktoken